requirements.txt
torch
torchvision
torchaudio
albumentations
opencv-python
matplotlib
scikit-learn
scikit-image
numpy
pyyaml
tqdm
pillow
pandas
onnx
onnxruntime
